Update HasNutrientRatios so .nutrient_ratios uses .get_nutrient_ratio.
Currently, the dependency is the other way around, which means we have to instantiate all of the nutrient ratios, just to get a single one.

This actually involved a reasonable overhaul of the HasSettableFlags tests, because by bad design, many of those were relying on the implementation of the .nutrient_ratios method. The nutrient ratio collection method is much more robustly tested after this work, because before and after tests were introduced to check that flag states were what they should have been, before the conflicts were collected.