Introduces caloric functionality into the model.
Brings in a new functionality class - HasNutrientMasses, and with it, methods to query the mass of a named nutrient on anything that implements HasNutrientMasses, and also a property to query the number of calories on anything that implements HasNutrientMasses.
Also updates HasNutrientRatios with an additional property to return the calories per gram.